Fix error where we free a buffer supplied as a "string" type.
A buffer can be supplied as either a "pointer" or a "string" type
argument. When a buffer was supplied as a string we inadvertantly
freed the underlying buffer.

This lead to the buffer content being replaced by arbitrary content.

Fixing this bug enables simplifying the API because now we don't
need a specific "convert buffer to string" method (although it
might still be useful) or maybe even the `ptr()` method..

Remind me to tell you the "funny story" about this one day...