Update to `godot-cpp` 9eceb16f0553884094d0f659461649be5d333866.
As of `godot-cpp` 9eceb16f0553884094d0f659461649be5d333866 a
number of fixes have been incorporated that mean `Ref<>`s are no
longer leaked and `PoolByteArray`/`PoolStringArray`s are also
no longer leaked when passed as arguments to GDNative functions.

The one(?) downside is that the `Ref<>` workaround that *was*
necessary...now causes a crash. *sigh*

Originally I thought `godot-cpp` latest (~3.2 but it's not on the
3.2 branch, yet) didn't work with Godot 3.1.x but it turned out
that the "mysterious load error" that occurred was Godot not
finding the test script! Apparently the "current directory" is
calculated different in Godot 3.1.x than 3.2.x but the error
returned wasn't stated as "file not found".

Because of what I thought was an incompatibility I figured out
how to get `godot-cpp` 3.1 to also not leak memory and while
that's not actually necessary now I thought I'd at least commit it,
at least temporarily, hence the conditionals. (Not yet committed.)

But the upshot is, this commit will probably be broken or leak
memory when compiled against `godot-cpp` 3.1 but should be good
for `godot-cpp` @ the mentioned ref.

More details: <https://github.com/godotengine/godot-cpp/issues/417>