Fix issue w/`TARGET_GODOT_CPP_3_2_LATEST` not defined everywhere.
Turned out that because `foreigner.h` isn't included everywhere
the code I thought was being used for `godot-cpp` 3.2 in some
places wasn't actually being used.

This pulls the define into a separate header file which seemed
simplest way of making this happen.

It might also be better to just require it to be a CLI macro
define or something but this will do for now.

This still leaves the `godot-cpp` 3.2 build broken because I now
need to update the code to be correct.